+++ app/components/timeseries/timeseries-service.js
@@ -198,9 +198,12 @@ angular.module('timeseries')
         return ts === undefined; // Break out early
       });
 
+      if (ts.reference_frame) {
+        graphTimeseries.unit += ' (' + ts.reference_frame + ')';
       if (ts) {
         graphTimeseries.parameter = ts.parameter || EMPTY;
         graphTimeseries.unit = ts.unit || EMPTY;
+        graphTimeseries.location = ts.location || EMPTY;
         if (ts.reference_frame) {
           graphTimeseries.unit += ' (' + ts.reference_frame + ')';
         }