+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
 
+- Add hover interaction to multiline graphs
+
 - Make the apps buttons a bit less ugly.
 
 - Store and restore timeseries state in dashboard. Refactored state.selected.